+++ editors/vscode/out/test/runTest.js
@@ -32,6 +32,7 @@ async function main() {
         await (0, test_electron_1.runTests)({
             extensionDevelopmentPath,
             extensionTestsPath,
+            launchArgs: [path.resolve(extensionDevelopmentPath, "../../demo")],
         });
     }
     catch (err) {